Refactoring of where test classes reside
