+++ seyren-core/src/test/java/com/seyren/core/service/MockCheck.java
@@ -1,4 +1,4 @@
-package com.seyren.mongo;
+package com.seyren.core.service;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
