+++ seyren-core/src/test/java/com/seyren/core/service/MockNotificationService.java
@@ -1,4 +1,4 @@
-package com.seyren.mongo;
+package com.seyren.core.service;
 
 import java.util.List;
