+++ seyren-core/src/test/java/com/seyren/core/service/MockSubscription.java
@@ -1,4 +1,4 @@
-package com.seyren.mongo;
+package com.seyren.core.service;
 
 import com.seyren.core.domain.Subscription;
